Time: 199 ms (68.93%) | Memory: 171.4 MB (87.06%) - LeetSync

+++ 3528-reach-end-of-array-with-max-score/reach-end-of-array-with-max-score.cpp
@@ -6,7 +6,7 @@ public:
         int j=0;
         for(int i=1;i<nums.size();i++){
             if(nums[i]>temp || i==nums.size()-1){
-                ans= ans + (long long)((i-j)*temp);
+                ans= ans + ((i-j)*temp);
                 temp=nums[i];
                 j=i;
             }
